feature: caching getPosts user's voting

+++ src/Services/Posts/Ytsoob.Services.Posts/Polls/Dtos/OptionDto.cs
@@ -5,5 +5,5 @@ public class OptionDto
     public long Id { get; set; }
     public string Title { get; set; } = default!;
     public long Count { get; set; }
-    public long Fiction { get; set; }
+    public decimal Fiction { get; set; }
 }
